pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
requests = "^2.32.3"
pandas = "2.1.4"
plotly = "^5.24.1"
nbformat = ">=4.2.0"
seaborn = "^0.13.2"
matplotlib = "^3.10.0"
tqdm = "^4.67.1"
scikit-learn = "1.5.2"
pyarrow = "^19.0.0"
xgboost = "^1.7.6"
lightgbm = "^4.5.0"
optuna = "^4.2.0"
python-dotenv = "^1.0.1"
hopsworks = {extras = ["python"], version = "^4.1.5"}
comet-ml = "^3.48.1"
streamlit = "^1.41.1"
geopandas = "^1.0.1"
deck = "^3.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
